# secp256k1: Introduce jacobian point struct.