… dcrec/secp256k1/bench_test.go
… dcrec/secp256k1/curve_test.go
… dcrec/secp256k1/genstatics.go