**This requires #2056**.

This introduces a new structure for passing around Jacobian coordinates and updates the code to make use of it.

This increases the readability of the code by reducing a lot of extra parameter clutter and makes it more obvious what type of coordinates the functions expect (Jacobian vs affine) as well as which parameters are for passing the results back out.  It also makes the internal code less cumbersome to work with.

The only reason it was originally split into coords was due to early versions of the Go compiler having issues with escape analysis leading to slower performance in performance-critical code.  However, that is no longer the case as can be seen by the benchmarks below which show everything within the margin of error in terms of performance.  A few reductions in allocations were achieved in the scalar multiplication path, but the focus of these changes are not optimization.

```
benchmark                         old ns/op    new ns/op    delta
------------------------------------------------------------------
BenchmarkAddJacobian              582          577          -0.86%
BenchmarkAddJacobianNotZOne       1091         1093         +0.18%
BenchmarkScalarBaseMult           46389        46681        +0.63%
BenchmarkScalarBaseMultJacobian   34437        34386        -0.15%
BenchmarkScalarBaseMultLarge      47012        47216        +0.43%
BenchmarkScalarMult               155915       156262       +0.22%
BenchmarkNAF                      1447         1372         -5.18%
BenchmarkSigVerify                215279       217324       +0.95%
BenchmarkFieldNormalize           21.3         21.5         +0.94%
BenchmarkNonceRFC6979             6402         6393         -0.14%

benchmark                         old allocs   new allocs   delta
------------------------------------------------------------------
 BenchmarkAddJacobian              0            0            +0.00%
 BenchmarkAddJacobianNotZOne       0            0            +0.00%
 BenchmarkScalarBaseMult           5            5            +0.00%
 BenchmarkScalarBaseMultJacobian   1            1            +0.00%
 BenchmarkScalarBaseMultLarge      8            8            +0.00%
 BenchmarkScalarMult               24           22           -8.33%
 BenchmarkNAF                      3            3            +0.00%
 BenchmarkSigVerify                64           58           -9.38%
 BenchmarkFieldNormalize           0            0            +0.00%
 BenchmarkNonceRFC6979             17           17           +0.00%

 benchmark                         old bytes    new bytes    delta
 ------------------------------------------------------------------
 BenchmarkAddJacobian              0            0            +0.00%
 BenchmarkAddJacobianNotZOne       0            0            +0.00%
 BenchmarkScalarBaseMult           224          224          +0.00%
 BenchmarkScalarBaseMultJacobian   32           32           +0.00%
 BenchmarkScalarBaseMultLarge      400          400          +0.00%
 BenchmarkScalarMult               1233         1137         -7.79%
 BenchmarkNAF                      128          128          +0.00%
 BenchmarkSigVerify                3138         2850         -9.18%
 BenchmarkFieldNormalize           0            0            +0.00%
 BenchmarkNonceRFC6979             976          976          +0.00%
```